Add Login Functionality
The functionality uses Facebook SDK to intialize authentication. The
code to initialize Facebook SDK is placed in the facebookAPI module. It
initializes the the SDK. After initialization, it checks the login
status. If loggedIn, it returns a promise with resolved value of true,
else it returns a resolved value of false. The promise value is then
passed to the React App initialization function as a prop. Which is then
passed to the <Login/> component.